Trait lignin::auto_safety::Align

pub trait Align<T: Vdom>: Vdom {
    fn align(self) -> T { ... }

    fn align_ref(&self) -> &T { ... }
}

Contextually thread-binds an instance, or not. Use only without qualification.

This trait acts as Into on and between variants of the same Vdom type, but without raising useless_conversion warnings.

See module documentation for when to use this trait and when it's unnecessary.

Provided methods

fn align(self) -> T

Contextually thread-binds an instance, or not. Use only without qualification.

fn align_ref(&self) -> &T

Contextually thread-binds a reference, or not. Use only without qualification.

Loading content...