Trait lignin::auto_safety::AutoSafe

pub trait AutoSafe<BoundVariant> where
    Self: Vdom + Align<BoundVariant>,
    BoundVariant: Vdom<ThreadSafety = ThreadBound>, {
#[must_use]    fn deanonymize(&self) -> BoundVariant { ... }
}

Deanonymize towards the general (ThreadBound) case. Used as -> impl AutoSafe<…>.

See module documentation for usage.

Provided methods

#[must_use]fn deanonymize(&self) -> BoundVariant

Deanonymize towards a compatible concrete type.

This method is by reference, so it will resolve with lower priority than the by-value method on Deanonymize.
Note that not all tooling will show the correct overload here, but the compiler knows which to pick.

Loading content...