Trait lignin::auto_safety::Deanonymize

pub trait Deanonymize<SafeVariant> where
    Self: Vdom + Send + Sync,
    SafeVariant: Vdom<ThreadSafety = ThreadSafe>, {
#[must_use]    fn deanonymize(self) -> SafeVariant { ... }
}

Deanonymize towards the special (ThreadSafe) case. This trait must be in scope for correct inference!

See module documentation for usage.

Provided methods

#[must_use]fn deanonymize(self) -> SafeVariant

Deanonymize towards a compatible concrete type.

This method is by value, so it will resolve with higher priority than the by-reference method on AutoSafe.
Note that not all tooling will show the correct overload here, but the compiler knows which to pick.

Loading content...