Module lignin::callback_registry

Callback registry plumbing, for renderers and app runners that support them and need to run indefinitely.

When not using this module directly, apps, if they enable the "callbacks" feature, run out of unique callback IDs after more than four billion total CallbackRegistration::new calls across all threads in a single run of the program. As such, you probably don't need to access this module, but if you do then it's available.

Structs

CallbackRef
Vdom A callback reference linked to a CallbackRegistration.

CallbackRegistration
A callback registration handle that should be held onto by the matching receiver R or a container with pin-projection towards that value.

Traits

ToRefThreadBoundFallback
Provides a fallback alternative implementation to CallbackRegistration::to_ref for use in macro frameworks.

Functions

registry_exhaustion
Indicates how exhausted the global callback registry is on a linear scale, with 0 indicating no or very low exhaustion and 255 indicating almost complete or complete exhaustion.

reset_callback_registry⚠
Tries to rewind the total callback registration counter to zero.

yet_more_unsafe_force_clear_callback_registry⚠
Clears the callback registry entirely and resets the total callback registration counter to zero.