Module lignin::web

Erasable web type stand-ins used as callback parameters.

structs in this module are only inhabited with the "callbacks" feature enabled.
Without it, they become uninhabited and are erased entirely at compile-time, so any code paths that depend on them can in turn be removed too.

Structs

Comment
Erasable stand-in for web_sys::Comment used as callback parameter.

Element
Erasable stand-in for web_sys::Element used as callback parameter.

Event
Erasable stand-in for web_sys::Event used as callback parameter.

HtmlElement
Erasable stand-in for web_sys::HtmlElement used as callback parameter.

SvgElement
Erasable stand-in for web_sys::SvgElement used as callback parameter.

Text
Erasable stand-in for web_sys::Text used as callback parameter.

Enums

DomRef
Used as DOM reference callback parameter. (Expand for implementation contract!)

Traits

Materialize
Convert a DOM stand-in to its web type value. This is a no-op with the "callbacks" feature and unreachable otherwise.