Trait lignin::web::Materialize

pub trait Materialize<T: Sized>: Sized + Sealed {
    fn materialize(self) -> T;
}

Convert a DOM stand-in to its web type value. This is a no-op with the "callbacks" feature and unreachable otherwise.

The extra trait is necessary because Into conflicts on T: From<T> and Option<T>: From<T>.

Warning:

Without the "callbacks" feature, the stand-ins in this module implement Materialize for any target type!
Make sure to check if your package compiles with this feature enables, most easily by requiring it in the [dev-dependencies] section of your Cargo.toml.

Required methods

fn materialize(self) -> T

Convert a DOM stand-in to its web type value. This is a no-op with the "callbacks" feature and unreachable otherwise.

Loading content...