Struct lignin::callback_registry::CallbackRegistration

pub struct CallbackRegistration<R, T> { /* fields omitted */ }

A callback registration handle that should be held onto by the matching receiver R or a container with pin-projection towards that value.

CallbackRegistration is !Unpin for convenience: A receiver correctly becomes !Unpin if it contains for example a Cell<Option<CallbackRegistration<R, T>>¹⁻².

impl<T: ?Sized> Unpin for Cell<T> where T: Unpin

impl<T> Unpin for Option<T> where T: Unpin

Implementations

impl<R, T> CallbackRegistration<R, T>

#[must_use]pub fn new(
receiver: Pin<&R>,
handler: fn(receiver: *const R, parameter: T)
) -> Self

Creates a new CallbackRegistration<R, T> with the given receiver and handler.

Safety

The receiver pointer given to handler may dangle unless receiver remains pinned until the created CallbackRegistration is dropped.

You can ensure this most easily by storing the latter in for example a Cell<Option<CallbackRegistration>> embedded in the receiver.

Dropping the CallbackRegistration instance prevents any further calls to handler through it.

#[must_use]pub fn to_ref_thread_bound(&self) -> CallbackRef<ThreadBound, T>

Creates a ThreadBound CallbackRef from this CallbackRegistration.

impl<R, T> CallbackRegistration<R, T> where
R: Sync,

#[must_use]pub fn to_ref(&self) -> CallbackRef<ThreadSafe, T>

Creates a ThreadSafe CallbackRef from this CallbackRegistration.

If you are developing a macro framework with ThreadSafety inference, see ToRefThreadBoundFallback for a way to overload this method appropriately.
(See also the warning there.)

For handwritten code or generated code with stricter thread-safety, please use .to_ref_thread_bound() instead whenever possible.